requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
contourpy==1.2.1
cycler==0.12.1
einops==0.7.0
fonttools==4.51.0
ftfy==6.2.0
grpcio==1.62.1
install==1.3.5
kiwisolver==1.4.5
lightning==2.2.2
lpips==0.1.4
Markdown==3.6
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
omegaconf==2.3.0
openai-clip==1.0.1
pillow==10.2.0
prettytable==3.10.0
protobuf==5.26.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.13.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
imageio==2.34.0
imageio-ffmpeg==0.4.9
torch-fidelity==0.3.0
